refactor cache(memoizer) and include tests around it
